Adds DoorButtonController inheritance
The previously named DoorButtonController and AdjacentDoorButtonController are now subclasses of the new DoorButtonController class
