+++ Assets/Scripts/Room/DoorSystem/MainDoorButtonController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8754cc8d6de361745979d40686c4683b
+guid: 912fd52b51203164ab72c00de79feabf
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
